chore: share RelayDnsPacket function code
